// src/product/product.pb.ts
export interface CreateProductRequest {
  name: string;
  sku: string;
  stock: number;
  price: number;
  category: string;
  description: string;
}

export interface CreateProductResponse {
  status: number;
  error: string[];
  id: number;
}

export interface FindOneData {
  id: number;
  name: string;
  sku: string;
  stock: number;
  price: number;
  category: string;
  description: string;
}

export interface FindOneRequest {
  id: number;
}

export interface FindOneResponse {
  status: number;
  error: string[];
  data: FindOneData | undefined;
}

export interface DecreaseStockRequest {
  id: number;
  orderId: number;
}

export interface DecreaseStockResponse {
  status: number;
  error: string[];
}

export interface UpdateProductRequest {
  productId: number;
  product: Product | undefined;
}

export interface UpdateProductResponse {
  status: number;
  error: string[];
  updatedProduct: Product | undefined;
}

export interface Product {
  id: number;
  name: string;
  sku: string;
  stock: number;
  price: number;
  category: string;
  description: string;
}

export interface FindByCategoryRequest {
  category: string;
}

export interface FindByCategoryResponse {
  status: number;
  error: string[];
  products: Product[];
}

export interface FindAllRequest {
}

export interface FindAllResponse {
  status: number;
  error: string[];
  products: Product[];
}

export interface FindByNameRequest {
  name: string;
}

export interface FindByNameResponse {
  status: number;
  error: string[];
  products: Product[];
}

export interface DeleteProductRequest {
  id: number;
}

export interface DeleteProductResponse {
  status: number;
  error: string[];
}

export const PRODUCT_PACKAGE_NAME = "product";

export interface ProductServiceClient {
  createProduct(request: CreateProductRequest): Observable<CreateProductResponse>;

  decreaseStock(request: DecreaseStockRequest): Observable<DecreaseStockResponse>;

  findOne(request: FindOneRequest): Observable<FindOneResponse>;

  findAll(request: FindAllRequest): Observable<FindAllResponse>;

  findByName(request: FindByNameRequest): Observable<FindByNameResponse>;

  findByCategory(request: FindByCategoryRequest): Observable<FindByCategoryResponse>;

  deleteProduct(request: DeleteProductRequest): Observable<DeleteProductResponse>;

  updateProduct(request: UpdateProductRequest): Observable<UpdateProductResponse>;
}
// ...
export const PRODUCT_SERVICE_NAME = "ProductService";